requirements.txt
opencv-python==4.6.0.66
numpy==1.23.4
Pillow==9.2.0
sklearn
webcolors